requirements.txt
moviepy==1.0.3
gtts==2.3.2
Pillow==10.1.0
selenium==4.14.0
webdriver-manager==4.0.1
numpy==1.24.3
requests==2.31.0
edge-tts==6.1.6
transformers==4.36.0
torch==2.1.0
scipy==1.11.0
huggingface-hub==0.19.0

# API Server
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0